Commit metodos create y update modification

+++ Proyecto Final/GraphQL/src/main/java/com/uptc/frm/graphql/controller/ClientController.java
@@ -1,7 +1,9 @@
 package com.uptc.frm.graphql.controller;
 
 import com.uptc.frm.graphql.jpa.models.Client;
+import com.uptc.frm.graphql.jpa.models.Repair;
 import com.uptc.frm.graphql.services.ClientService;
+import com.uptc.frm.graphql.services.RepairService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
@@ -14,6 +16,8 @@ import java.util.List;
 public class ClientController {
     @Autowired
     private ClientService clientService;
+    @Autowired
+    private RepairService repairService;
     @QueryMapping(name = "findClientById")
     public Client findClientById(@Argument(name = "idClientC") Integer idClient){
         return clientService.findById(idClient);
@@ -34,8 +38,13 @@ public class ClientController {
     public String deleteClient(@Argument Integer deleteClient){
         Client client = clientService.findById(deleteClient);
         if (client != null) {
-            clientService.deleteClient(deleteClient);
-            return ("El cliente con id: " + deleteClient + " se elimino con exito");
+            List<Repair> repairs = repairService.findByNumberIdClient(deleteClient);
+            if (repairs.isEmpty()) {
+                clientService.deleteClient(deleteClient);
+                return ("El cliente con id: " + deleteClient + " se elimino con exito");
+            } else{
+                return ("El cliente con id: " + deleteClient + " tiene "+ repairs.size()+" reparaciones, por lo tanto no se puede eliminar");
+            }
         }else {
             return ("El cliente con id: " + deleteClient + " no existe");
         }
